+++ poem/src/web/compress.rs
@@ -8,6 +8,7 @@ use tokio::io::{AsyncRead, BufReader};
 
 use crate::{
     http::{header, HeaderValue},
+    web::CompressionLevel,
     Body, IntoResponse, Response,
 };
 
@@ -15,11 +16,10 @@ use crate::{
 #[cfg_attr(docsrs, doc(cfg(feature = "compression")))]
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum CompressionAlgo {
-    // /// brotli
-    // BR,
+    /// brotli
+    BR,
     /// deflate
     DEFLATE,
-
     /// gzip
     GZIP,
 }
@@ -29,7 +29,7 @@ impl FromStr for CompressionAlgo {
 
     fn from_str(s: &str) -> std::prelude::rust_2015::Result<Self, Self::Err> {
         Ok(match s {
-            // "br" => CompressionAlgo::BR,
+            "br" => CompressionAlgo::BR,
             "deflate" => CompressionAlgo::DEFLATE,
             "gzip" => CompressionAlgo::GZIP,
             _ => return Err(()),
@@ -38,9 +38,10 @@ impl FromStr for CompressionAlgo {
 }
 
 impl CompressionAlgo {
+    #[inline]
     pub(crate) fn as_str(&self) -> &'static str {
         match self {
-            // CompressionAlgo::BR => "br",
+            CompressionAlgo::BR => "br",
             CompressionAlgo::DEFLATE => "deflate",
             CompressionAlgo::GZIP => "gzip",
         }
@@ -49,18 +50,27 @@ impl CompressionAlgo {
     pub(crate) fn compress<'a>(
         &self,
         reader: impl AsyncRead + Send + Unpin + 'a,
+        level: Option<CompressionLevel>,
     ) -> Pin<Box<dyn AsyncRead + Send + 'a>> {
         match self {
-            // CompressionAlgo::BR =>
-            // Box::pin(async_compression::tokio::bufread::BrotliEncoder::new(
-            //     BufReader::new(reader),
-            // )),
-            CompressionAlgo::DEFLATE => Box::pin(
-                async_compression::tokio::bufread::DeflateEncoder::new(BufReader::new(reader)),
+            CompressionAlgo::BR => Box::pin(
+                async_compression::tokio::bufread::BrotliEncoder::with_quality(
+                    BufReader::new(reader),
+                    level.unwrap_or(CompressionLevel::Fastest),
+                ),
             ),
-            CompressionAlgo::GZIP => Box::pin(async_compression::tokio::bufread::GzipEncoder::new(
-                BufReader::new(reader),
-            )),
+            CompressionAlgo::DEFLATE => Box::pin(
+                async_compression::tokio::bufread::DeflateEncoder::with_quality(
+                    BufReader::new(reader),
+                    level.unwrap_or(CompressionLevel::Default),
+                ),
+            ),
+            CompressionAlgo::GZIP => Box::pin(
+                async_compression::tokio::bufread::GzipEncoder::with_quality(
+                    BufReader::new(reader),
+                    level.unwrap_or(CompressionLevel::Default),
+                ),
+            ),
         }
     }
 
@@ -69,10 +79,9 @@ impl CompressionAlgo {
         reader: impl AsyncRead + Send + Unpin + 'a,
     ) -> Pin<Box<dyn AsyncRead + Send + 'a>> {
         match self {
-            // CompressionAlgo::BR =>
-            // Box::pin(async_compression::tokio::bufread::BrotliDecoder::new(
-            //     BufReader::new(reader),
-            // )),
+            CompressionAlgo::BR => Box::pin(async_compression::tokio::bufread::BrotliDecoder::new(
+                BufReader::new(reader),
+            )),
             CompressionAlgo::DEFLATE => Box::pin(
                 async_compression::tokio::bufread::DeflateDecoder::new(BufReader::new(reader)),
             ),
@@ -109,12 +118,27 @@ impl Display for CompressionAlgo {
 pub struct Compress<T> {
     inner: T,
     algo: CompressionAlgo,
+    level: Option<CompressionLevel>,
 }
 
 impl<T> Compress<T> {
-    /// /// Create a compressed response using the specified algorithm.
+    /// Create a compressed response using the specified algorithm.
     pub fn new(inner: T, algo: CompressionAlgo) -> Self {
-        Self { inner, algo }
+        Self {
+            inner,
+            algo,
+            level: None,
+        }
+    }
+
+    /// Specify the compression level
+    #[must_use]
+    #[inline]
+    pub fn with_quality(self, level: CompressionLevel) -> Self {
+        Self {
+            level: Some(level),
+            ..self
+        }
     }
 }
 
@@ -130,7 +154,7 @@ impl<T: IntoResponse> IntoResponse for Compress<T> {
         resp.headers_mut().remove(header::CONTENT_LENGTH);
 
         resp.set_body(Body::from_async_read(
-            self.algo.compress(body.into_async_read()),
+            self.algo.compress(body.into_async_read(), self.level),
         ));
         resp
     }
@@ -176,7 +200,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_compress() {
-        // test_algo(CompressionAlgo::BR).await;
+        test_algo(CompressionAlgo::BR).await;
         test_algo(CompressionAlgo::DEFLATE).await;
         test_algo(CompressionAlgo::GZIP).await;
     }