+++ poem/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.48] 2022-10-20
+
+- Re-enable the `brotli` compression algorithm.
+- Add `Compress::with_quality` and `Compression::with_quality` methods.
+
 # [1.3.47] 2022-10-19
 
 - Fixes [#346](https://github.com/poem-web/poem/issues/346) [#395](https://github.com/poem-web/poem/issues/395)